Lazy SMP!
Lazy SMP is actually slower for me in time to depth than the serial version, but I have a theory that while it is slower, it's actually playing better and is actually looking deeper than it shows. I think that it is actually looking around twice as deep as it shows (because other threads are searching at a deeper depth than the main thread), which causes there to be less pruning on the main thread. As a result, this causes it to be slower, but stronger because it is technically looking deeper. This is all a theory though, I don't know if that's actually what's going on.