Improvements (see description)
Improved performance of pawn move generation. Now, all non-pinned pawn moves are generated in one go. Pinned pawns are still generated one by one since we have to mask by the pinned piece mobility.

11% improvement in search performance.

Also, fixed a bug with pins that existed previously. Before, pieces could be pinned to our king through an opponent piece. This is fixed though.

The node count for a search on the start position with depth 8 has suspiciously changed though. It's not because of the bug fix, because the node count was still different before fixing the bug. Therefore, it is possible there is a bug with the batch pawn move generation, but I have not found any issues with it yet. Move generation is passing perft tests (without en passant, since it is not implemented yet), and the search PV is still the same, so I'm not entirely sure what is causing this. Either way, I'm just going to ignore this until a larger issue with it comes up.