Improved evaluation function
Added a king attack heuristic to the king safety evaluation. The king attack heuristic still has a long way to go, but I feel like the engine is starting to create more attacks on the enemy king now.

Also, added a separate king endgame table, and made the evaluations interpolate between the middlegame and endgame.