Working legality checker
We are now able to determine if a move is pseudo-legal (and legal).

Also, we now check that the hash move is both pseudo-legal and legal before searching it in case of rare hash key collisions (see https://www.chessprogramming.org/Hash_Move and https://www.chessprogramming.org/Transposition_Table#KeyCollisions). Surprisingly, this extra check did not slow down the search much.

The main reason I added this is because I am planning to add in the killer move heuristic, which will definitely require legality tests.