Killer heuristic
Weirdly enough, according to testing, it seems that both the history heuristic and killer heuristic make the search performance significantly worse, and that pure transposition table move ordering is the best.

Not entirely sure why this is (could be a bug?), but I am keeping both the history and killer heuristic in for now.