Removed transposition table locking
I recall seeing somewhere that a Stockfish dev said that not locking the transposition table actually lead to a stronger engine despite concurrent writes. Not sure if this is true, but at the moment, this engine does not have multithreading enabled, so the locking isn't really an issue right now.

If this does become an issue in the future, I can bring the locking code back.

Also, removing the transposition table locking alone gave a ~7% increase in search performance.