Added a basic NNUE accumulator
This increases the nodes per second by about 5x.

The accumulator still has a lot of work left to do. A lot of things can be vectorized in the accumulator with SIMD. Additionally, the weights for the first hidden layer can be reorganized in memory to improve cache locality (this would also enable us to vectorize the `Accumulator::add` and `Accumulator::remove` functions). Finally, we should store the state of the accumulator in `MakeMoveInfo` to help alleviate any possible floating point errors that may have accumulated.