Made the network smaller
This network is not very mature, it was only trained for 3 epochs. Yet it still manages to destroy the previous network (previous one was trained on a full 10 epochs).

Nodes per second roughly tripled due to the smaller network. I guess this hints at the fact that nodes per second matters more than the quality of the network. As such, aggressive network quantization seems like a must-have. Will implement that soon.

Gains 135.8 +/- 43.7 Elo.