Refactored Board's internal usage of MakeMoveFlags
This makes it a lot easier to implement internal Board methods correctly, since now, everything in the Board class that doesn't need to be incrementally updated when unmaking a move (i.e. everything that is stored in the MakeMoveInfo struct) can just be marked as a MakeMoveFlag that is mutually exclusive to the MakeMoveFlags::Unmake flag.

Besides all the technical changes, I'm pretty sure that this also improves performance of the engine a bit since I just realized that we were incrementally updating the hash during unmakes, which is totally unnecessary and useless since it would get overwritten by the old hash from MakeMoveInfo anyways.

No functional change.